requirements.txt
pandas==1.5.3
pymongo==4.4.0
python-dotenv==1.0.0
streamlit==1.22.0
plotly==5.17.0
nltk
folium
streamlit-folium